+++ test/fixture/mock-generator/trait-self-type/builder.php
@@ -1,0 +1,7 @@
+<?php
+
+use Eloquent\Phony\Test\TestTraitWithSelfType;
+
+$builder = $factory->create(TestTraitWithSelfType::class);
+
+return $builder->named('MockGeneratorTraitSelfType');